+++ examples/with-nextjs-and-clerk-auth/src/components/MoniteComponents.tsx
@@ -105,8 +105,9 @@ export const MoniteProvider = ({
       locale={{
         code: i18n.locale,
         messages: {
-          Payables: 'Purchases',
+          Payables: 'Bill Pay',
           Counterpart: 'Customer',
+          Sales: 'Invoicing',
         },
       }}
     >