+++ packages/sdk-react/mui-styles.d.ts
@@ -7,6 +7,7 @@ import { type MonitePayableStatusChipProps } from '@/components/payables/Payable
 import { type MoniteInvoiceRecurrenceIterationStatusChipProps } from '@/components/receivables/InvoiceRecurrenceIterationStatusChip/InvoiceRecurrenceIterationStatusChip';
 import { type MoniteInvoiceRecurrenceStatusChipProps } from '@/components/receivables/InvoiceRecurrenceStatusChip/InvoiceRecurrenceStatusChip';
 import { type MoniteInvoiceStatusChipProps } from '@/components/receivables/InvoiceStatusChip/InvoiceStatusChip';
+import { type MoniteIconWrapperProps } from '@/ui/iconWrapper/IconWrapper';
 import { type MoniteTablePaginationProps } from '@/ui/table/TablePagination';
 import {
   ComponentsOverrides,
@@ -43,6 +44,7 @@ declare module '@mui/material/styles' {
     MoniteCounterpartStatusChip: 'root';
     MoniteApprovalStatusChip: 'root';
     MonitePayableTable: 'never';
+    MoniteReceivablesTable: 'never';
   }
 
   /**
@@ -59,6 +61,8 @@ declare module '@mui/material/styles' {
     MoniteCounterpartStatusChip: Partial<MoniteCounterpartStatusChipProps>;
     MonitePayableTable: Partial<MonitePayableTableProps>;
     MoniteApprovalStatusChip: Partial<MoniteApprovalStatusChipProps>;
+    MoniteReceivablesTable: Partial<MoniteReceivablesTableProps>;
+    MoniteIconWrapper: Partial<MoniteIconWrapperProps>;
   }
 
   interface MoniteOptions {
@@ -81,5 +85,7 @@ declare module '@mui/material/styles' {
     MoniteInvoiceRecurrenceIterationStatusChip?: ComponentType<'MoniteInvoiceRecurrenceIterationStatusChip'>;
     MoniteCounterpartStatusChip?: ComponentType<'MoniteCounterpartStatusChip'>;
     MoniteApprovalStatusChip?: ComponentType<'MoniteApprovalStatusChip'>;
+    MoniteReceivablesTable?: ComponentType<'MoniteReceivablesTable'>;
+    MoniteIconWrapper?: ComponentType<'MoniteIconWrapper'>;
   }
 }