+++ packages/sdk-react/src/components/payables/PayablesTable/components/PayablesTableAction.tsx
@@ -1,4 +1,5 @@
 import { components } from '@/api';
+import { usePaymentHandler } from '@/components/payables/PayablesTable/hooks/usePaymentHandler';
 import { useIsActionAllowed } from '@/core/queries/usePermissions';
 import { t } from '@lingui/macro';
 import { useLingui } from '@lingui/react';
@@ -6,12 +7,6 @@ import { Button } from '@mui/material';
 
 interface PayablesTableActionProps {
   payable: components['schemas']['PayableResponseSchema'];
-
-  /**
-   * The event handler for the pay action
-   *
-   * @param id - The identifier of the row to perform the pay action on, a string.
-   */
   onPay?: (id: string) => void;
 }
 
@@ -26,26 +21,33 @@ export const PayablesTableAction = ({
     entityUserId: payable.was_created_by_user_id,
   });
 
-  if (isPayAllowed && payable.status === 'waiting_to_be_paid') {
-    return (
-      <Button
-        variant="outlined"
-        size="small"
-        onClick={(e) => {
-          /**
-           * We have to stop propagation to disable
-           *  `onRowClick` callback when the user
-           *  clicks on the `Pay` button
-           */
-          e.stopPropagation();
+  const { handlePay, modalComponent, isPaymentLinkAvailable } =
+    usePaymentHandler(payable.id);
 
-          onPay?.(payable.id);
-        }}
-      >
-        {t(i18n)`Pay`}
-      </Button>
-    );
+  if (!isPayAllowed) {
+    return null;
   }
 
-  return null;
+  if (isPayAllowed && payable.status === 'waiting_to_be_paid') {
+    return (
+      <>
+        {isPaymentLinkAvailable ? (
+          <Button
+            variant="outlined"
+            size="small"
+            onClick={(e) => {
+              e.stopPropagation();
+              onPay?.(payable.id);
+              handlePay();
+            }}
+          >
+            {t(i18n)`Pay`}
+          </Button>
+        ) : (
+          <p>{t(i18n)`Payment not allowed`}</p>
+        )}
+        {modalComponent}
+      </>
+    );
+  }
 };