+++ packages/sdk-react/src/components/products/ProductDetails/ProductEditForm/ProductEditForm.tsx
@@ -8,6 +8,7 @@ import { useMoniteContext } from '@/core/context/MoniteContext';
 import { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
 import { useCurrencies } from '@/core/hooks';
 import { CenteredContentBox } from '@/ui/box';
+import { IconWrapper } from '@/ui/iconWrapper';
 import { LoadingPage } from '@/ui/loadingPage';
 import { t } from '@lingui/macro';
 import { useLingui } from '@lingui/react';
@@ -20,7 +21,6 @@ import {
   DialogContent,
   Divider,
   Grid,
-  IconButton,
   Stack,
   Typography,
 } from '@mui/material';
@@ -98,13 +98,13 @@ const ProductEditFormBase = (props: IProductEditFormProps) => {
           </Grid>
           <Grid item xs={1}>
             {dialogContext?.isDialogContent && (
-              <IconButton
+              <IconWrapper
                 aria-label={t(i18n)`Edit Product Close`}
                 onClick={dialogContext.onClose}
                 color="inherit"
               >
                 <CloseIcon />
-              </IconButton>
+              </IconWrapper>
             )}
           </Grid>
         </Grid>
@@ -175,13 +175,13 @@ const ProductEditFormBase = (props: IProductEditFormProps) => {
         </Grid>
         <Grid item xs={1}>
           {dialogContext?.isDialogContent && (
-            <IconButton
+            <IconWrapper
               aria-label={t(i18n)`Edit Product Close`}
               onClick={dialogContext.onClose}
               color="inherit"
             >
               <CloseIcon />
-            </IconButton>
+            </IconWrapper>
           )}
         </Grid>
       </Grid>