+++ packages/sdk-react/src/components/receivables/InvoiceDetails/CreateReceivable/sections/components/ReminderForm/OverdueReminderForm.tsx
@@ -6,6 +6,7 @@ import { components } from '@/api';
 import { RHFTextField } from '@/components/RHF/RHFTextField';
 import { useMoniteContext } from '@/core/context/MoniteContext';
 import { getAPIErrorMessage } from '@/core/utils/getAPIErrorMessage';
+import { IconWrapper } from '@/ui/iconWrapper';
 import { LoadingPage } from '@/ui/loadingPage';
 import { NotFound } from '@/ui/notFound';
 import { yupResolver } from '@hookform/resolvers/yup';
@@ -19,7 +20,6 @@ import {
   DialogContent,
   DialogActions,
   Divider,
-  IconButton,
   Stack,
   Typography,
   InputLabel,
@@ -177,7 +177,7 @@ const CreateOverdueReminderComponent = ({
               : t(i18n)`Create “Overdue” reminder`}
           </Typography>
           {props.onClose && (
-            <IconButton
+            <IconWrapper
               edge="start"
               color="inherit"
               onClick={(event) => {
@@ -187,7 +187,7 @@ const CreateOverdueReminderComponent = ({
               aria-label={t(i18n)`Close reminder's creation`}
             >
               <CloseIcon />
-            </IconButton>
+            </IconWrapper>
           )}
         </Box>
       </DialogTitle>